[BE] Refactor: entities 폴더 이름 models로 변경, property 이름 camelCase로 변경
- entities 폴더 이름 models로 변경, 기존 MongoDB user model 주석 처리
- property 이름 snake_case -> camelCase로 변경
