+++ server/src/models/User.ts
@@ -1,9 +1,31 @@
-import mongoose, { mongo } from 'mongoose';
+import { Entity, PrimaryGeneratedColumn, Column } from 'typeorm';
 
-const userSchema = new mongoose.Schema({
-    id: {type: String, required: true, unique: true},
-    pw: {type: String, required: true, trim: true},
-});
-userSchema.index({id: 1});
+@Entity()
+class User {
+    @PrimaryGeneratedColumn()
+    id!: number;
 
-export default mongoose.model('User', userSchema);
+    @Column()
+    email: string;
+
+    @Column()
+    password: string;
+
+    @Column()
+    name: string;
+
+    @Column()
+    imageUrl: string;
+}
+
+export default User;
+
+// import mongoose, { mongo } from 'mongoose';
+
+// const userSchema = new mongoose.Schema({
+//     id: {type: String, required: true, unique: true},
+//     pw: {type: String, required: true, trim: true},
+// });
+// userSchema.index({id: 1});
+
+// export default mongoose.model('User', userSchema);
