+++ server/src/routes/user.ts
@@ -1,13 +1,13 @@
-import express from 'express';
-import User from '../models/User';
-const router = express.Router();
+// import express from 'express';
+// import User from '../models/User';
+// const router = express.Router();
 
-router.get('/:id',  (req, res, next) => {
-    const id = req.params.id;
-    console.log(User);
-    User.find({})
-    .then((user) => res.send(user))
-    .catch(e =>  res.send(e));
-})
+// router.get('/:id',  (req, res, next) => {
+//     const id = req.params.id;
+//     console.log(User);
+//     User.find({})
+//     .then((user) => res.send(user))
+//     .catch(e =>  res.send(e));
+// })
 
-export default router;
+// export default router;
